Split remaining Ast transformers
