+++ src/ast/transform/SemanticChecker.cpp
@@ -47,6 +47,7 @@
 #include "ast/analysis/SCCGraph.h"
 #include "ast/analysis/Type.h"
 #include "ast/analysis/TypeEnvironment.h"
+#include "ast/transform/GroundedTermsChecker.h"
 #include "utility/ContainerUtil.h"
 #include "utility/FunctionalUtil.h"
 #include "utility/MiscUtil.h"
@@ -1188,50 +1189,6 @@ void AstSemanticCheckerImpl::checkNamespaces() {
     }
 }
 
-bool AstExecutionPlanChecker::transform(AstTranslationUnit& translationUnit) {
-    auto* relationSchedule = translationUnit.getAnalysis<RelationScheduleAnalysis>();
-    auto* recursiveClauses = translationUnit.getAnalysis<RecursiveClausesAnalysis>();
-    auto&& report = translationUnit.getErrorReport();
-
-    for (const RelationScheduleAnalysisStep& step : relationSchedule->schedule()) {
-        const std::set<const AstRelation*>& scc = step.computed();
-        for (const AstRelation* rel : scc) {
-            for (const AstClause* clause : getClauses(*translationUnit.getProgram(), *rel)) {
-                if (!recursiveClauses->recursive(clause)) {
-                    continue;
-                }
-                if (clause->getExecutionPlan() == nullptr) {
-                    continue;
-                }
-                int version = 0;
-                for (const auto* atom : getBodyLiterals<AstAtom>(*clause)) {
-                    if (scc.count(getAtomRelation(atom, translationUnit.getProgram())) != 0u) {
-                        version++;
-                    }
-                }
-                int maxVersion = -1;
-                for (auto const& cur : clause->getExecutionPlan()->getOrders()) {
-                    maxVersion = std::max(cur.first, maxVersion);
-                }
-
-                if (version <= maxVersion) {
-                    for (const auto& cur : clause->getExecutionPlan()->getOrders()) {
-                        if (cur.first >= version) {
-                            report.addDiagnostic(Diagnostic(Diagnostic::Type::ERROR,
-                                    DiagnosticMessage(
-                                            "execution plan for version " + std::to_string(cur.first),
-                                            cur.second->getSrcLoc()),
-                                    {DiagnosticMessage("only versions 0.." + std::to_string(version - 1) +
-                                                       " permitted")}));
-                        }
-                    }
-                }
-            }
-        }
-    }
-    return false;
-}
-
 void TypeChecker::visitAtom(const AstAtom& atom) {
     auto relation = getAtomRelation(&atom, &program);
     if (relation == nullptr) {
@@ -1492,33 +1449,6 @@ void TypeChecker::visitAggregator(const AstAggregator& aggregator) {
     if (!isOfKind(aggregatorType, opType)) {
         report.addError("Couldn't assign types to the aggregator", aggregator.getSrcLoc());
     }
-}
-
-void GroundedTermsChecker::verify(AstTranslationUnit& translationUnit) {
-    auto&& program = *translationUnit.getProgram();
-    auto&& report = translationUnit.getErrorReport();
-
-    // -- check grounded variables and records --
-    visitDepthFirst(program.getClauses(), [&](const AstClause& clause) {
-        if (isFact(clause)) return;  // only interested in rules
-
-        auto isGrounded = getGroundedTerms(translationUnit, clause);
-
-        std::set<std::string> reportedVars;
-        // all terms in head need to be grounded
-        for (auto&& cur : getVariables(clause)) {
-            if (!isGrounded[cur] && reportedVars.insert(cur->getName()).second) {
-                report.addError("Ungrounded variable " + cur->getName(), cur->getSrcLoc());
-            }
-        }
-
-        // all records need to be grounded
-        for (auto&& cur : getRecords(clause)) {
-            if (!isGrounded[cur]) {
-                report.addError("Ungrounded record", cur->getSrcLoc());
-            }
-        }
-    });
 }
 
 }  // end of namespace souffle
