+++ src/ast/transform/Transformer.cpp
@@ -36,18 +36,4 @@ bool AstTransformer::apply(AstTranslationUnit& translationUnit) {
     return changed;
 }
 
-bool MetaTransformer::applySubtransformer(AstTranslationUnit& translationUnit, AstTransformer* transformer) {
-    auto start = std::chrono::high_resolution_clock::now();
-    bool changed = transformer->apply(translationUnit);
-    auto end = std::chrono::high_resolution_clock::now();
-
-    if (verbose && (dynamic_cast<MetaTransformer*>(transformer) == nullptr)) {
-        std::string changedString = changed ? "changed" : "unchanged";
-        std::cout << transformer->getName() << " time: " << std::chrono::duration<double>(end - start).count()
-                  << "sec [" << changedString << "]" << std::endl;
-    }
-
-    return changed;
-}
-
 }  // end of namespace souffle
